backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
redis==5.0.4
aioredis==2.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
anthropic==0.28.0
ccxt==4.3.42
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
aiohttp==3.9.5
httpx==0.27.0
cryptography==42.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyotp==2.9.0
qrcode==7.4.2
apscheduler==3.10.4
python-multipart==0.0.9
openpyxl==3.1.4
reportlab==4.2.0
loguru==0.7.2
tenacity==8.3.0
prometheus-client==0.20.0
